Reland^2 "Don't use AOT code for native methods for java debuggable runtime"
This reverts commit 570ade8a6600d368a9e24b64cfa0a1907929166a.

Reason for revert: Relanding after a fix for failures. The original cl breaks the invariant that we would always use AOT code for native methods if there is AOT code. This invariant is necessary to get the header when walking the stack. This CL fixes it by not relying on the invariant but instead tagging the sp to differentiate between JIT and AOT code in debuggable runtimes. Non-debuggable runtimes still have the invariant.

Change-Id: I5141281f04202d41988021d53bfe30a48bc4db9c